+++ database/migrations/2018_10_12_144531_create_user_favorite_products_table.php
@@ -14,10 +14,10 @@ class CreateUserFavoriteProductsTable extends Migration
     public function up()
     {
         Schema::create('user_favorite_products', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
         });
